requirements.txt
requests==2.27.1
python-dotenv==0.20.0
web3==5.28.0